+++ End.hpp
@@ -5,7 +5,7 @@
 class End : public ISection
 {
 public:
-	End(glm::vec3 end, int direction);
+	End(glm::vec3 end, float direction);
 	int update(float& delta, IModel* model, int section);
 
 private:
